Feature/add gql subscritpion

+++ apps/frontend/src/app/chatroom/chatroom-list/chatroom-list.component.ts
@@ -1,11 +1,13 @@
 import {Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import {ChatroomService} from "../chatroom.service";
-import {Subscription} from "rxjs";
-import {ChatroomsQuery} from "../../graphql/generated";
+import {combineLatest, Subscription} from "rxjs";
+import {ChatroomsQuery, WhoAmIQuery} from "../../graphql/generated";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {TranslateService} from "@ngx-translate/core";
 import {Router} from "@angular/router";
 import {MatSelectionList} from "@angular/material/list";
+import {ConfirmationDialogComponent} from "../../lib/confirmation-dialog/confirmation-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'mindmap-chatroom-list',
@@ -15,23 +17,26 @@ import {MatSelectionList} from "@angular/material/list";
 export class ChatroomListComponent implements OnDestroy, OnInit {
 
   chatrooms: ChatroomsQuery["chatrooms"] = [];
+  user !: WhoAmIQuery["whoAmI"];
   @ViewChild('roomList') roomList!: MatSelectionList;
   subscription: Subscription[] = []
 
   constructor(private chatroomService: ChatroomService,
               private _snackBar: MatSnackBar,
               private t: TranslateService,
-              private router: Router) {
-  }
-
-  ngOnDestroy(): void {
-    this.subscription.forEach(sub => sub.unsubscribe());
+              private router: Router,
+              private dialog: MatDialog,) {
   }
 
   ngOnInit(): void {
-    this.subscription.push(this.chatroomService.getAllChatrooms().subscribe(chatrooms => {
-      this.chatrooms = chatrooms;
-    }));
+
+    this.subscription.push(
+      combineLatest([
+        this.chatroomService.getAllChatrooms(),
+        this.chatroomService.whoAmI()]).subscribe(([chatrooms, user]) => {
+        this.chatrooms = chatrooms;
+        this.user = user;
+      }));
     this.subscription.push(this.chatroomService.chatroom$.subscribe((data) => {
       if (!data) return;
       this.chatrooms = [...data];
@@ -39,18 +44,42 @@ export class ChatroomListComponent implements OnDestroy, OnInit {
   }
 
   deleteRoom(id: string) {
-    this.subscription.push(this.chatroomService.deleteChatroom(id).subscribe({
-      next: () => {
-        this._snackBar.open(this.t.instant("DELETE_CHATROOM_MESSAGE_SUCCESSES"), 'X', {
-          horizontalPosition: 'center',
-          verticalPosition: 'top',
-        });
-        this.router.navigate(['chatrooms']);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        title: 'CHATROOMS.DELETE_CHATROOM',
+        description: 'CHATROOMS.DELETE_CHATROOM_WARNING',
       },
-    }));
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result === true) {
+        this.subscription.push(this.chatroomService.deleteChatroom(id).subscribe({
+          next: () => {
+            this._snackBar.open(this.t.instant("DELETE_CHATROOM_MESSAGE_SUCCESSES"), 'X', {
+              horizontalPosition: 'center',
+              verticalPosition: 'top',
+            });
+            this.router.navigate(['chatrooms']);
+          }
+        }));
+      }
+    })
   }
 
-  navigateToMessageView(id?: string) {
-   this.router.navigate(['chatrooms', id]);
+  navigateToMessageView(room: ChatroomsQuery["chatrooms"][number]) {
+    const include = room.users.some(user => user.id === this.user.id);
+    if (room.type === 'PUBLIC' || include) {
+      this.router.navigate(['chatrooms', room.id]);
+      return;
+    }
+    this._snackBar.open(this.t.instant("PRIVATE_CHATROOM"), 'X', {
+      horizontalPosition: 'center',
+      verticalPosition: 'top',
+    });
   }
+
+  ngOnDestroy(): void {
+    this.subscription.forEach(sub => sub.unsubscribe());
+  }
+
 }
